Fix some issues in file processing task

+++ src/main/java/edu/mum/tm/repository/FileProcessingRepository.java
@@ -10,6 +10,6 @@ import java.util.List;
 
 
 @Repository
-public interface FileProcessingRepository extends CrudRepository<FileProcessing, Long>  {
+public interface FileProcessingRepository extends CrudRepository<FileProcessing, Integer>  {
 
 }
